+++ CHANGES.rst
@@ -9,7 +9,7 @@
 
 - Created inverse non-linearity schema. [#213]
 
-- Add ``archive.destinations`` introspection function to collect all 
+- Add ``archive.destinations`` introspection function to collect all
   ``archive_catalog.destination`` values. [#218]